updating

+++ src/main/java/es/etg/daw/dawes/java/rest/restfull/productos/infraestructure/web/rest/ProductoController.java
@@ -6,6 +6,7 @@ import org.springframework.web.bind.annotation.RestController;
 import es.etg.daw.dawes.java.rest.restfull.productos.application.command.CreateProductoCommand;
 import es.etg.daw.dawes.java.rest.restfull.productos.application.command.EditProductoCommand;
 import es.etg.daw.dawes.java.rest.restfull.productos.application.service.CreateProductoService;
+import es.etg.daw.dawes.java.rest.restfull.productos.application.service.DeleteProductoService;
 import es.etg.daw.dawes.java.rest.restfull.productos.application.service.EditProductoService;
 import es.etg.daw.dawes.java.rest.restfull.productos.application.service.FindProductoService;
 import es.etg.daw.dawes.java.rest.restfull.productos.domain.model.Producto;
@@ -67,7 +68,7 @@ public class ProductoController {
 
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<?>  deleteProducto(@PathVariable Long id) {
+    public ResponseEntity<?>  deleteProducto(@PathVariable int id) {
         DeleteProductoService.delete(id);
         return ResponseEntity.noContent().build(); //Devpñvemos una respuesta vacía.
     }
